Initial Commit
This commit is very much in a beta stage. Not all the pages are created or completed. Same 4 or 5 pictures used throughout. No real design work is done. Jus the initial work on most pages are done. This will help me see how it reacts on different screens and on an actual web server